+++ src/lib.rs
@@ -972,7 +972,10 @@ impl<Fs: FileSystem> ResolverGeneric<Fs> {
                 }
             }
             if should_stop {
-                return Err(ResolveError::NotFound(specifier.to_string()));
+                return Err(ResolveError::MatchedAliasNotFound(
+                    specifier.to_string(),
+                    alias_key.to_string(),
+                ));
             }
         }
         Ok(None)
@@ -1011,7 +1014,9 @@ impl<Fs: FileSystem> ResolverGeneric<Fs> {
             *should_stop = true;
             ctx.with_fully_specified(false);
             return match self.require(cached_path, new_specifier.as_ref(), ctx) {
-                Err(ResolveError::NotFound(_)) => Ok(None),
+                Err(ResolveError::NotFound(_) | ResolveError::MatchedAliasNotFound(_, _)) => {
+                    Ok(None)
+                }
                 Ok(path) => return Ok(Some(path)),
                 Err(err) => return Err(err),
             };
@@ -1042,11 +1047,12 @@ impl<Fs: FileSystem> ResolverGeneric<Fs> {
         else {
             return Ok(None);
         };
-        let path = cached_path.path().with_extension("");
-        let path = path.as_os_str();
+        let path = cached_path.path();
+        let Some(filename) = path.file_name() else { return Ok(None) };
+        let path_without_extension = path.with_extension("");
         ctx.with_fully_specified(true);
         for extension in extensions {
-            let mut path_with_extension = path.to_os_string();
+            let mut path_with_extension = path_without_extension.clone().into_os_string();
             path_with_extension.reserve_exact(extension.len());
             path_with_extension.push(extension);
             let cached_path = self.cache.value(Path::new(&path_with_extension));
@@ -1060,7 +1066,16 @@ impl<Fs: FileSystem> ResolverGeneric<Fs> {
                 return Ok(Some(path));
             }
         }
-        Err(ResolveError::ExtensionAlias(cached_path.to_path_buf()))
+        // Create a meaningful error message.
+        let dir = path.parent().unwrap().to_path_buf();
+        let filename_without_extension = Path::new(filename).with_extension("");
+        let filename_without_extension = filename_without_extension.to_string_lossy();
+        let files = extensions
+            .iter()
+            .map(|ext| format!("{filename_without_extension}{ext}"))
+            .collect::<Vec<_>>()
+            .join(",");
+        Err(ResolveError::ExtensionAlias(filename.to_string_lossy().to_string(), files, dir))
     }
 
     /// enhanced-resolve: RootsPlugin
@@ -1283,8 +1298,9 @@ impl<Fs: FileSystem> ResolverGeneric<Fs> {
             let mut has_dot = false;
             let mut without_dot = false;
             for key in map.keys() {
-                has_dot = has_dot || key.starts_with(|s| s == '.' || s == '#');
-                without_dot = without_dot || !key.starts_with(|s| s == '.' || s == '#');
+                let starts_with_dot_or_hash = key.starts_with(['.', '#']);
+                has_dot = has_dot || starts_with_dot_or_hash;
+                without_dot = without_dot || !starts_with_dot_or_hash;
                 if has_dot && without_dot {
                     return Err(ResolveError::InvalidPackageConfig(
                         package_url.join("package.json"),