+++ src/tests/alias.rs
@@ -231,7 +231,10 @@ fn all_alias_values_are_not_found() {
         ..ResolveOptions::default()
     });
     let resolution = resolver.resolve(&f, "m1/a.js");
-    assert_eq!(resolution, Err(ResolveError::NotFound("m1/a.js".to_string())));
+    assert_eq!(
+        resolution,
+        Err(ResolveError::MatchedAliasNotFound("m1/a.js".to_string(), "m1".to_string(),))
+    );
 }
 
 #[test]